Implemented a solution for parallel frequency
I also tried this without calling Frequency from the ConcurrentFrequency function and directly looping over the string. I thought this would be faster because 1) we are avoiding two for loops and 2) we are not calling Frequency anymore. However, it was at least twice as slow. This is where the performance effects of mutexes and benefits of sync.Map() come up. However, I can't think of how to use sync.Map() without mutexes without changing the signature of Frequency(). And I can't change that because the test calls that function.